refactor : websocket,redis -> domain 폴더로 이동
