+++ src/main/java/com/codeit/otboo/domain/dm/service/DmServiceImpl.java
@@ -15,10 +15,10 @@ import com.codeit.otboo.domain.dm.dto.DirectMessageDto;
 import com.codeit.otboo.domain.dm.dto.DirectMessageDtoCursorResponse;
 import com.codeit.otboo.domain.dm.entity.Dm;
 import com.codeit.otboo.domain.dm.mapper.DirectMessageMapper;
-import com.codeit.otboo.domain.dm.redis.RedisPublisher;
+import com.codeit.otboo.domain.redis.RedisPublisher;
 import com.codeit.otboo.domain.dm.repository.DmRepository;
 import com.codeit.otboo.domain.dm.util.DmKeyUtil;
-import com.codeit.otboo.domain.dm.websocket.NewDmEvent;
+import com.codeit.otboo.domain.websocket.listener.NewDmEvent;
 import com.codeit.otboo.domain.notification.dto.NotificationDto;
 import com.codeit.otboo.domain.notification.entity.NotificationLevel;
 import com.codeit.otboo.domain.notification.service.NotificationService;
