+++ src/test/java/com/codeit/otboo/domain/redis/RedisPublisherTest.java
@@ -8,8 +8,6 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.data.redis.core.RedisTemplate;
 
-import com.codeit.otboo.domain.dm.redis.RedisPublisher;
-
 @ExtendWith(MockitoExtension.class)
 public class RedisPublisherTest {
